@@ -30,18 +30,13 @@ Future<bool> workoutExists(String title) async {
 }
 
 Future<Workout> loadWorkout(String title) async {
-  try {
-    final file = await _loadWorkoutFile(title);
+  final file = await _loadWorkoutFile(title);
+  var contents = await file.readAsString();
 
-    var contents = await file.readAsString();
-
-    return Workout.fromRawJson(contents);
-  } on Exception {
-    return null;
-  }
+  return Workout.fromRawJson(contents);
 }
 
-void deleteWorkout(String title) async {
+Future<void> deleteWorkout(String title) async {
   try {
     final file = await _loadWorkoutFile(title);
     file.delete();
